sdk/pyproject.toml
[project]
name = "langperf"
version = "0.2.0"
description = "LangPerf SDK — tracer for agent trajectories"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-exporter-otlp-proto-http>=1.28",
    "openinference-instrumentation-openai>=0.1.18",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.11",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["langperf*"]

[tool.setuptools.package-data]
# Ship the py.typed marker so mypy/pyright consume our inline type hints.
langperf = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
# Start conservative — only rules that pass today.
# E/W = pycodestyle, F = pyflakes, I = isort, UP = pyupgrade,
# B = flake8-bugbear (real bugs).
select = ["E", "W", "F", "I", "UP", "B"]
ignore = [
  "E501",  # handled by formatter
  "B008",  # FastAPI Depends() in defaults is idiomatic
  # --- Baseline ignores: existing violations, tighten later ---
  "I001",  # un-sorted imports — don't churn existing import blocks
  "UP007", # `Optional[X]` vs `X | None` — stylistic
  "UP037", # quoted type annotations — stylistic
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["B011"]          # assert False in tests is fine
"**/alembic/versions/**" = ["E402", "F401", "I"]  # migrations are frozen

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
# Lax to start — tighten later. No strict mode.
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = false
ignore_missing_imports = true
# Don't check third-party stubs we don't control
exclude = ["alembic/versions/", "tests/"]
